+++ smc-monitoring/HISTORY.rst
@@ -7,4 +7,13 @@ Release History
 ++++++++++++++++++
 
 - Initial release of SMC Monitoring package
+- Requires SMC 6.3.0, 6.3.1
+
+1.1.0 (2018-02-02)
+++++++++++++++++++
+
+- Requires SMC >= 6.3.2
+- ActiveAlertQuery implemented for interfacing with Alerts (requires SMC >= 6.3.3). Currently read-only
+- BlacklistQuery modified to use generic log fields for all queries except fetch_as_element
+